Replace Python publishing workflow and add unit test workflow
Removed the Python publishing workflow (python-publish.yml) as it was outdated and replaced it with a more efficient and autonomous workflow (pypi_publish.yml). This change automates the process of publishing a Python package to PyPI whenever a new tagged release is pushed.

Also, added a new workflow (tests.yml) that automatically runs unit tests across different Python versions and operating systems whenever a pull request is made to the master branch. This ensures the package maintains high code quality and compatibility across different platforms.