"Updated dependencies and dropped contourpy package"
In the given diff, several modifications were implemented towards optimizing the dependencies pinned in our poetry.lock file. Specifically, dependencies "exceptiongroup" and "typing_extensions" were added to cater for any python version less than 3.11. On top of that, the version of the "cycler" package was updated to "0.12.1" from "0.12.0" for features or fixes that come with the new version. Modifications were also made to the 'gevent' and 'matplotlib' packages. Lastly, the 'contourpy' package was entirely removed from the dependencies list due to either it being no longer used in our project, or its features being replaced by other packages.